refactor: split validateQuery into separate module
This fixes https://github.com/confuser/graphql-constraint-directive/issues/156
